@jreback 2015-09-21
ping when green (and tests are moved)