@chris-b1 2015-09-21
@jreback - moved the `Period` test as suggested, travis is green. Thanks.